// src/components/Table/types.ts
export type Row = Record<string, string> & {
	_id?: string | number;
	id: string | number;
};

// Manage Columns
export type TComponentType =
	| 'link'
	| 'date'
	| 'dateTime'
	| 'duration'
	| 'status'
	| 'time';

export interface TManageColumnProps {
	row: Row;
	rowDetail?: () => React.ReactNode;
	columns: IColumns;
	changeColumns: React.Dispatch<React.SetStateAction<IColumns>>;
	closePopover: () => void;
	manageColumnButtonProps?: Object;
}

export interface TManageColumnsItemProps {
	id: string | number;
	index: number;
	columnName: string;
	row?: Row;
	onChange: () => void;
	isVisible: boolean | undefined;
	isMandatory?: boolean;
}
// ...
export interface IColumn {
	type?: TComponentType;
	key: string;
	options?: Record<string, string>;
	field?: string;
	isVisible?: boolean;
	label: string;
	sortable?: boolean;
	colSpan?: number | string;
	render: () => React.ReactNode;
	isMandatory?: boolean;
}

export type IColumns = IColumn[];